+++ changelog.md
@@ -4,6 +4,8 @@ Codion Change Log
 ## 0.18.84
 ### is.codion.common.reactive
 - ValueCollection.removeAll() wildcard added to mirror addAll().
+### is.codion.framework.domain
+- DefaultForeignKeyConditions.valueMap() bug fixed, no longer uses toMap(), due to nulls. Now rejects non-present keys.
 
 ## 0.18.83
 ### is.codion